govguardian_app/backend/src/main/webapp/app/entities/gov-official-content/route/gov-official-content-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IGovOfficialContent } from '../gov-official-content.model';
import { GovOfficialContentService } from '../service/gov-official-content.service';

import govOfficialContentResolve from './gov-official-content-routing-resolve.service';

describe('GovOfficialContent routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let service: GovOfficialContentService;
  let resultGovOfficialContent: IGovOfficialContent | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    service = TestBed.inject(GovOfficialContentService);
    resultGovOfficialContent = undefined;
  });

  describe('resolve', () => {
    it('should return IGovOfficialContent returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        govOfficialContentResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultGovOfficialContent = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultGovOfficialContent).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      TestBed.runInInjectionContext(() => {
        govOfficialContentResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultGovOfficialContent = result;
          },
        });
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultGovOfficialContent).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IGovOfficialContent>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        govOfficialContentResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultGovOfficialContent = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultGovOfficialContent).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
